feat: tests refactoring
